Fix for new default RRA policy in ganglia
The new default for the RRD database samplying frequency take 20MB per node
before this change it was 300KB per node this can make the tmpfs reserved for
the RRD full for big cluster.

I crafted this policy which is between the new default and the old and it takes
about 3M per node.